+++ R/go_enrich_net.R
@@ -47,7 +47,7 @@ go_enrich_net <- function(go_anno,
 										 net_layout = "circle",
 										 net_circular = TRUE,
 										 low_color = "#ff0000aa",
-										 high_color = "#008888aa"
+										 high_color = "#008800aa"
 										){
 	# -> 2. Data Parameters
 	# padjust_method <- "fdr"
@@ -164,15 +164,21 @@ go_enrich_net <- function(go_anno,
 	p <- cnetplot(
 		x = enrich_results,
 		showCategory = category_num,
-		foldChange = deg_list,
+		color.params = list(foldChange = deg_list,
+												edge = TRUE),
+		# foldChange = deg_list,
+		# colorEdge = TRUE,
+		cex.params = list(category_node = 1,
+											gene_node = 1,
+											category_label = cateLabelScale,
+											gene_label = geneLabelScale),
+		# cex_category = 1,
+		# cex_gene = 1,
+		# cex_label_category = cateLabelScale,
+		# cex_label_gene = geneLabelScale,
 		layout = net_layout, # 'star', 'circle', 'gem', 'dh', 'graphopt', 'grid', 'mds', 'randomly', 'fr', 'kk', 'drl' or 'lgl'.
-		colorEdge = TRUE,
 		circular = net_circular,
 		node_label = "all",
-		cex_category = 1,
-		cex_gene = 1,
-		cex_label_category = cateLabelScale,
-		cex_label_gene = geneLabelScale,
 		# color_category = "#ff0000",
 		# color_gene = "#008000",
 		shadowtext = "all"